renamed the file to EthereumConsoleManager
