+++ examples/18-eth-private-network/EthereumConsoleManager.py
